fix: Use exception when requests are too large.
